+++ Sources/Classes/Client/Plugins/RSAdvertisementIdPlugin.swift
@@ -22,14 +22,28 @@ class RSAdvertisingIdPlugin: RSPlatformPlugin {
             context[keyPath: "device.advertisingId"] = advertisingId
             context[keyPath: "device.adTrackingEnabled"] = true
             workingMessage.context = context
+            client?.updateContext(context)
         }
         return workingMessage
     }
 }
 
 extension RSClient {
+    /**
+     API for setting identifier under context.device.advertisingId.
+     - Parameters:
+        - advertisingId: IDFA value
+     # Example #
+     ```
+     client.setAdvertisingId("sample_device_token")
+     ```
+     */
     @objc
     public func setAdvertisingId(_ advertisingId: String) {
+        guard advertisingId.isNotEmpty else {
+            log(message: "advertisingId can not be empty", logLevel: .warning)
+            return
+        }
         if let advertisingIdPlugin = self.find(pluginType: RSAdvertisingIdPlugin.self) {
             advertisingIdPlugin.advertisingId = advertisingId
         } else {