+++ Sources/Classes/Domain/Models/Attributes/RSServerConfig.swift
@@ -8,7 +8,7 @@
 
 import Foundation
 
-@objc open class RSServerConfig: NSObject, Codable {
+public class RSServerConfig: NSObject, Codable {
     
     struct Source: Codable {
         private let _id: String?
@@ -137,5 +137,20 @@ extension RSServerConfig {
             }
         }
         return nil
+    }
+    
+    private func getConfig<T: Codable>(forKey key: String) -> T? {
+        var result: T?
+        if let destination = getDestination(by: key) {
+            guard let config = destination.config?.dictionaryValue else { return nil }
+            if let jsonData = try? JSONSerialization.data(withJSONObject: config) {
+                result = try? JSONDecoder().decode(T.self, from: jsonData)
+            }
+        }
+        return result
+    }
+    
+    public func getConfig<T: Codable>(forPlugin plugin: RSDestinationPlugin) -> T? {
+        return getConfig(forKey: plugin.key)
     }
 }