+++ Sources/Classes/Domain/Models/RSOption.swift
@@ -22,6 +22,14 @@ open class RSOption: NSObject {
     
     @objc
     public func putExternalId(_ type: String, withId id: String) {
+        guard type.isNotEmpty else {
+            RSClient.rsLog(message: "ExternalId type can not be empty", logLevel: .warning)
+            return
+        }
+        guard id.isNotEmpty else {
+            RSClient.rsLog(message: "External id can not be empty", logLevel: .warning)
+            return
+        }
         if externalIds == nil {
             externalIds = [[String: String]]()
         }
@@ -37,16 +45,22 @@ open class RSOption: NSObject {
     
     @objc
     public func putIntegration(_ type: String, isEnabled enabled: Bool) {
+        guard type.isNotEmpty else {
+            RSClient.rsLog(message: "Integration type can not be empty", logLevel: .warning)
+            return
+        }
         integrations?[type] = enabled
     }
         
     @objc
-    public func putCustomContext(_ context: [String: Any]?, withKey key: String?) {
+    public func putCustomContext(_ context: [String: Any], withKey key: String) {
+        guard key.isNotEmpty else {
+            RSClient.rsLog(message: "CustomContext key can not be empty", logLevel: .warning)
+            return
+        }
         if customContexts == nil {
             customContexts = [String: Any]()
         }
-        if let key = key, let context = context {
-            customContexts?[key] = context
-        }
+        customContexts?[key] = context
     }
 }